IOS-4215: Create separate entity for WalletModel info providers
