+++ Tangem/Modules/Main/MultiWalletMainContent/Views/TokenList/TokenItemViewState+WalletModelState.swift
@@ -1,21 +1,14 @@
 //
-//  TokenItemViewState.swift
+//  TokenItemViewState+WalletModelState.swift
 //  Tangem
 //
-//  Created by Andrew Son on 08/08/23.
+//  Created by Andrew Son on 11/08/23.
 //  Copyright © 2023 Tangem AG. All rights reserved.
 //
 
 import Foundation
 
-enum TokenItemViewState {
-    case notLoaded
-    case noDerivation
-    case loading
-    case loaded
-    case noAccount(message: String)
-    case networkError(Error)
-
+extension TokenItemViewState {
     init(walletModelState: WalletModel.State) {
         switch walletModelState {
         case .created:
