+++ Tangem/UIComponents/TokenItemView/TokenItemViewState.swift
@@ -1,0 +1,18 @@
+//
+//  TokenItemViewState.swift
+//  Tangem
+//
+//  Created by Andrew Son on 08/08/23.
+//  Copyright © 2023 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+enum TokenItemViewState {
+    case notLoaded
+    case noDerivation
+    case loading
+    case loaded
+    case noAccount(message: String)
+    case networkError(Error)
+}
